Pass Field Info via Context
When resolving a field, it may be necessary to understand what the
'alias' is that the field is being resolved as. Whilst this shouldn't
generally impact the execution, understanding the difference between the
name of the field and its alias is sometimes useful.

This API is *NOT* intended to be stable. It is provided only as an
experiment, to aid some migration work.